refactor: standardize types in COREMOD_arith sub-modules (Phase 3) (#325)
- Replaced legacy 'long' with explicit width 'uint64_t' and 'int64_t'
- Standardized image coordinate parameters in image_crop to 'int64_t'
- Updated function prototypes in execute_arith and COREMOD_arith
- Ensured consistent SIMD optimization potential across routines

+++ src/coremods/COREMOD_arith/execute_arith.h
@@ -9,6 +9,6 @@
  */
 
 imageID
-arith_make_slopexy(const char *ID_name, long l1, long l2, double sx, double sy);
+arith_make_slopexy(const char *ID_name, uint32_t l1, uint32_t l2, double sx, double sy);
 
 int execute_arith(const char *cmd1);
